Add files via upload
Line 80 there is an example of how to download a video from a specified link and what the element looks like.

mediaUrls( ) function now returns all urls which lead to meeting media.

Line 49 you can see which CSS selectors find these elements. Then a for loop stores all of the urls into a list and returns it.